Move stuff to new file cluster_tree.cpp
